Rename Log::new -> new_with_bytes, add new_with_entries.
The `new` constructor was sometimes awkward to use because it's
hard to know how many entries the log will end up having since it
takes log size in bytes as an argument.
We add a `new_with_entries` constructor which creates a log with
`X` entries. We rename the `new` constructor to `new_with_bytes`
for consistent naming.